+++ gruntfile.js
@@ -111,12 +111,7 @@ module.exports = function (grunt) {
 
     bump: {
       options: {
-        files: [
-          'package.json',
-          'package-lock.json',
-          'bower.json',
-          'iframeResizer.jquery.json'
-        ],
+        files: ['package.json', 'package-lock.json', 'bower.json'],
         updateConfigs: ['pkg'],
         commit: true,
         commitMessage: 'Release v%VERSION%',