+++ js/iframeResizer.contentWindow.js
@@ -9,7 +9,7 @@
  */
 
 // eslint-disable-next-line sonarjs/cognitive-complexity, no-shadow-restricted-names
-;(function(undefined) {
+;(function (undefined) {
   if (typeof window === 'undefined') return // don't run for server side render
 
   var autoResize = true,
@@ -32,6 +32,7 @@
     interval = 32,
     intervalTimer = null,
     logging = false,
+    mouseEvents = false,
     msgID = '[iFrameSizer]', // Must match host page msg ID
     msgIdLen = msgID.length,
     myID = '',
@@ -53,17 +54,17 @@
     widthCalcModeDefault = 'scroll',
     widthCalcMode = widthCalcModeDefault,
     win = window,
-    onMessage = function() {
+    onMessage = function () {
       warn('onMessage function not defined')
     },
-    onReady = function() {},
-    onPageInfo = function() {},
+    onReady = function () {},
+    onPageInfo = function () {},
     customCalcMethods = {
-      height: function() {
+      height: function () {
         warn('Custom height calculation function not defined')
         return document.documentElement.offsetHeight
       },
-      width: function() {
+      width: function () {
         warn('Custom width calculation function not defined')
         return document.body.scrollWidth
       }
@@ -78,7 +79,7 @@
       {},
       {
         passive: {
-          get: function() {
+          get: function () {
             passiveSupported = true
           }
         }
@@ -109,7 +110,7 @@
       result,
       timeout = null,
       previous = 0,
-      later = function() {
+      later = function () {
         previous = getNow()
         timeout = null
         result = func.apply(context, args)
@@ -119,7 +120,7 @@
         }
       }
 
-    return function() {
+    return function () {
       var now = getNow()
 
       if (!previous) {
@@ -154,7 +155,7 @@
 
   var getNow =
     Date.now ||
-    function() {
+    function () {
       /* istanbul ignore next */ // Not testable in PhantonJS
       return new Date().getTime()
     }
@@ -179,7 +180,7 @@
 
   function init() {
     readDataFromParent()
-    log('Initialising iFrame (' + location.href + ')')
+    log('Initialising iFrame (' + window.location.href + ')')
     readDataFromPage()
     setMargin()
     setBodyStyle('background', bodyBackground)
@@ -189,6 +190,7 @@
     checkWidthMode()
     stopInfiniteResizingOfIFrame()
     setupPublicMethods()
+    setupMouseEvents()
     startEventListeners()
     inPageLinks = setupInPageLinks()
     sendSize('init', 'Init message from host page')
@@ -216,6 +218,7 @@
     inPageLinks.enable = undefined !== data[12] ? strBool(data[12]) : false
     resizeFrom = undefined !== data[13] ? data[13] : resizeFrom
     widthCalcMode = undefined !== data[14] ? data[14] : widthCalcMode
+    mouseEvents = undefined !== data[15] ? Boolean(data[15]) : mouseEvents
   }
 
   function depricate(key) {
@@ -311,7 +314,7 @@
 
   function manageTriggerEvent(options) {
     var listener = {
-      add: function(eventName) {
+      add: function (eventName) {
         function handleEvent() {
           sendSize(options.eventName, options.eventType)
         }
@@ -320,7 +323,7 @@
 
         addEventListener(window, eventName, handleEvent, { passive: true })
       },
-      remove: function(eventName) {
+      remove: function (eventName) {
         var handleEvent = eventHandlersByName[eventName]
         delete eventHandlersByName[eventName]
 
@@ -582,8 +585,11 @@
     }
 
     function checkLocationHash() {
-      if ('' !== location.hash && '#' !== location.hash) {
-        findTarget(location.href)
+      var hash = window.location.hash
+      var href = window.location.href
+
+      if ('' !== hash && '#' !== hash) {
+        findTarget(href)
       }
     }
 
@@ -641,6 +647,22 @@
     }
   }
 
+  function setupMouseEvents() {
+    if (mouseEvents !== true) return
+
+    function sendMouse(e) {
+      sendMsg(0, 0, e.type, e.screenY + ':' + e.screenX)
+    }
+
+    function addMouseListener(evt, name) {
+      log('Add event listener: ' + name)
+      addEventListener(window.document, evt, sendMouse)
+    }
+
+    addMouseListener('mouseenter', 'Mouse Enter')
+    addMouseListener('mouseleave', 'Mouse Leave')
+  }
+
   function setupPublicMethods() {
     log('Enable public methods')
 
@@ -671,7 +693,7 @@
           onPageInfo = callback
           sendMsg(0, 0, 'pageInfo')
         } else {
-          onPageInfo = function() {}
+          onPageInfo = function () {}
           sendMsg(0, 0, 'pageInfoStop')
         }
       },
@@ -731,7 +753,7 @@
   function initInterval() {
     if (0 !== interval) {
       log('setInterval: ' + interval + 'ms')
-      intervalTimer = setInterval(function() {
+      intervalTimer = setInterval(function () {
         sendSize('interval', 'setInterval: ' + interval)
       }, Math.abs(interval))
     }
@@ -773,7 +795,7 @@
 
     function imageEventTriggered(event, type, typeDesc) {
       removeImageLoadListener(event.target)
-      sendSize(type, typeDesc + ': ' + event.target.src, undefined, undefined)
+      sendSize(type, typeDesc + ': ' + event.target.src)
     }
 
     function imageLoaded(event) {
@@ -819,7 +841,7 @@
       observer = createMutationObserver()
 
     return {
-      disconnect: function() {
+      disconnect: function () {
         if ('disconnect' in observer) {
           log('Disconnect body MutationObserver')
           observer.disconnect()
@@ -894,12 +916,12 @@
     return maxVal
   }
 
-  function getAllMeasurements(dimention) {
+  function getAllMeasurements(dimensions) {
     return [
-      dimention.bodyOffset(),
-      dimention.bodyScroll(),
-      dimention.documentElementOffset(),
-      dimention.documentElementScroll()
+      dimensions.bodyOffset(),
+      dimensions.bodyScroll(),
+      dimensions.documentElementOffset(),
+      dimensions.documentElementScroll()
     ]
   }
 
@@ -911,7 +933,7 @@
 
     var elements = document.querySelectorAll('[' + tag + ']')
 
-    if (0 === elements.length) noTaggedElementsFound()
+    if (elements.length === 0) noTaggedElementsFound()
 
     return getMaxElement(side, elements)
   }
@@ -929,7 +951,7 @@
         )
       },
 
-      offset: function() {
+      offset: function () {
         return getHeight.bodyOffset() // Backwards compatability
       },
 
@@ -1113,7 +1135,7 @@
       log('Trigger event lock on')
     }
     clearTimeout(triggerLockedTimer)
-    triggerLockedTimer = setTimeout(function() {
+    triggerLockedTimer = setTimeout(function () {
       triggerLocked = false
       log('Trigger event lock off')
       log('--')
@@ -1175,7 +1197,7 @@
 
         init()
         firstRun = false
-        setTimeout(function() {
+        setTimeout(function () {
           initLock = false
         }, eventCancelTimer)
       },