+++ js/iframeResizer.js
@@ -9,7 +9,7 @@
  */
 
 // eslint-disable-next-line sonarjs/cognitive-complexity, no-shadow-restricted-names
-;(function(undefined) {
+;(function (undefined) {
   if (typeof window === 'undefined') return // don't run for server side render
 
   var count = 0,
@@ -46,6 +46,7 @@
       maxWidth: Infinity,
       minHeight: 0,
       minWidth: 0,
+      mouseEvents: true,
       resizeFrom: 'parent',
       scrolling: false,
       sizeHeight: true,
@@ -53,16 +54,18 @@
       warningTimeout: 5000,
       tolerance: 0,
       widthCalculationMethod: 'scroll',
-      onClose: function() {
+      onClose: function () {
         return true
       },
-      onClosed: function() {},
-      onInit: function() {},
-      onMessage: function() {
+      onClosed: function () {},
+      onInit: function () {},
+      onMessage: function () {
         warn('onMessage function not defined')
       },
-      onResized: function() {},
-      onScroll: function() {
+      onMouseEnter: function () {},
+      onMouseLeave: function () {},
+      onResized: function () {},
+      onScroll: function () {
         return true
       }
     }
@@ -98,7 +101,7 @@
       // Firefox extension content-scripts have a globalThis object that is not the same as window.
       // Binding `requestAnimationFrame` to window allows the function to work and prevents errors
       // being thrown when run in that context, and should be a no-op in every other context.
-      requestAnimationFrame = requestAnimationFrame.bind(window);
+      requestAnimationFrame = requestAnimationFrame.bind(window)
     }
   }
 
@@ -106,6 +109,7 @@
     var retStr = 'Host page: ' + iframeId
 
     if (window.top !== window.self) {
+      // eslint-disable-next-line unicorn/prefer-ternary
       if (window.parentIFrame && window.parentIFrame.getId) {
         retStr = window.parentIFrame.getId() + ': ' + iframeId
       } else {
@@ -159,8 +163,8 @@
 
     function processMsg() {
       var data = msg.substr(msgIdLen).split(':')
-      var height = data[1] ? parseInt(data[1], 10) : 0;
-      var iframe = settings[data[0]] && settings[data[0]].iframe;
+      var height = data[1] ? parseInt(data[1], 10) : 0
+      var iframe = settings[data[0]] && settings[data[0]].iframe
       var compStyle = getComputedStyle(iframe)
 
       return {
@@ -174,19 +178,25 @@
 
     function getPaddingEnds(compStyle) {
       if (compStyle.boxSizing !== 'border-box') {
-        return 0;
+        return 0
       }
       var top = compStyle.paddingTop ? parseInt(compStyle.paddingTop, 10) : 0
-      var bot = compStyle.paddingBottom ? parseInt(compStyle.paddingBottom, 10) : 0
+      var bot = compStyle.paddingBottom
+        ? parseInt(compStyle.paddingBottom, 10)
+        : 0
       return top + bot
     }
 
     function getBorderEnds(compStyle) {
       if (compStyle.boxSizing !== 'border-box') {
-        return 0;
+        return 0
       }
-      var top = compStyle.borderTopWidth ? parseInt(compStyle.borderTopWidth, 10) : 0
-      var bot = compStyle.borderBottomWidth ? parseInt(compStyle.borderBottomWidth, 10) : 0
+      var top = compStyle.borderTopWidth
+        ? parseInt(compStyle.borderTopWidth, 10)
+        : 0
+      var bot = compStyle.borderBottomWidth
+        ? parseInt(compStyle.borderBottomWidth, 10)
+        : 0
       return top + bot
     }
 
@@ -291,10 +301,12 @@
           msgBody +
           '}'
       )
+
       on('onMessage', {
         iframe: messageData.iframe,
         message: JSON.parse(msgBody)
       })
+
       log(iframeId, '--')
     }
 
@@ -341,7 +353,7 @@
           }
         }
 
-        ;['scroll', 'resize'].forEach(function(evt) {
+        ;['scroll', 'resize'].forEach(function (evt) {
           log(id, type + evt + ' listener for sendPageInfo')
           func(window, evt, sendPageInfo)
         })
@@ -500,6 +512,30 @@
       }
     }
 
+    function onMouse(event) {
+      var mousePos = {}
+
+      if (Number(messageData.width) === 0 && Number(messageData.height) === 0) {
+        var data = getMsgBody(9).split(':')
+        mousePos = {
+          x: data[1],
+          y: data[0]
+        }
+      } else {
+        mousePos = {
+          x: messageData.width,
+          y: messageData.height
+        }
+      }
+
+      on(event, {
+        iframe: messageData.iframe,
+        screenX: Number(mousePos.x),
+        screenY: Number(mousePos.y),
+        type: messageData.type
+      })
+    }
+
     function on(funcName, val) {
       return chkEvent(iframeId, funcName, val)
     }
@@ -514,6 +550,14 @@
 
         case 'message':
           forwardMsgFromIFrame(getMsgBody(6))
+          break
+
+        case 'mouseenter':
+          onMouse('onMouseEnter')
+          break
+
+        case 'mouseleave':
+          onMouse('onMouseLeave')
           break
 
         case 'autoResize':
@@ -554,7 +598,19 @@
           break
 
         default:
-          resizeIFrame()
+          if (
+            Number(messageData.width) === 0 &&
+            Number(messageData.height) === 0
+          ) {
+            warn(
+              'Unsupported message received (' +
+                messageData.type +
+                '), this is likely due to the iframe containing a later ' +
+                'version of iframe-resizer than the parent page'
+            )
+          } else {
+            resizeIFrame()
+          }
       }
     }
 
@@ -764,9 +820,12 @@
 
   function syncResize(func, messageData, doNotSync) {
     /* istanbul ignore if */ // Not testable in PhantomJS
-    if (doNotSync !== messageData.type && requestAnimationFrame &&
-        // including check for jasmine because had trouble getting spy to work in unit test using requestAnimationFrame
-        !window.jasmine) {
+    if (
+      doNotSync !== messageData.type &&
+      requestAnimationFrame &&
+      // including check for jasmine because had trouble getting spy to work in unit test using requestAnimationFrame
+      !window.jasmine
+    ) {
       log(messageData.id, 'Requesting animation frame')
       requestAnimationFrame(func)
     } else {
@@ -873,7 +932,9 @@
       ':' +
       settings[iframeId].resizeFrom +
       ':' +
-      settings[iframeId].widthCalculationMethod
+      settings[iframeId].widthCalculationMethod +
+      ':' +
+      settings[iframeId].mouseEvents
     )
   }
 
@@ -1016,7 +1077,7 @@
             settings[iframeId].iframe
           ),
 
-          moveToAnchor: function(anchor) {
+          moveToAnchor: function (anchor) {
             trigger(
               'Move to anchor',
               'moveToAnchor:' + anchor,
@@ -1025,7 +1086,7 @@
             )
           },
 
-          sendMessage: function(message) {
+          sendMessage: function (message) {
             message = JSON.stringify(message)
             trigger(
               'Send Message',
@@ -1052,10 +1113,10 @@
           return
         }
 
-        var destroyObserver = new MutationObserver(function(mutations) {
-          mutations.forEach(function(mutation) {
+        var destroyObserver = new MutationObserver(function (mutations) {
+          mutations.forEach(function (mutation) {
             var removedNodes = Array.prototype.slice.call(mutation.removedNodes) // Transform NodeList into an Array
-            removedNodes.forEach(function(removedNode) {
+            removedNodes.forEach(function (removedNode) {
               if (removedNode === iframe) {
                 closeIFrame(iframe)
               }
@@ -1097,7 +1158,10 @@
     }
 
     function getTargetOrigin(remoteHost) {
-      return '' === remoteHost || null !== remoteHost.match(/^(about:blank|javascript:|file:\/\/)/) ? '*' : remoteHost
+      return '' === remoteHost ||
+        null !== remoteHost.match(/^(about:blank|javascript:|file:\/\/)/)
+        ? '*'
+        : remoteHost
     }
 
     function depricate(key) {
@@ -1124,12 +1188,7 @@
       settings[iframeId] = {
         firstRun: true,
         iframe: iframe,
-        remoteHost:
-          iframe.src &&
-          iframe.src
-            .split('/')
-            .slice(0, 3)
-            .join('/')
+        remoteHost: iframe.src && iframe.src.split('/').slice(0, 3).join('/')
       }
 
       checkOptions(options)
@@ -1164,7 +1223,7 @@
 
   function debouce(fn, time) {
     if (null === timer) {
-      timer = setTimeout(function() {
+      timer = setTimeout(function () {
         timer = null
         fn()
       }, time)
@@ -1174,7 +1233,7 @@
   var frameTimer = {}
   function debounceFrameEvents(fn, time, frameId) {
     if (!frameTimer[frameId]) {
-      frameTimer[frameId] = setTimeout(function() {
+      frameTimer[frameId] = setTimeout(function () {
         frameTimer[frameId] = null
         fn()
       }, time)
@@ -1212,7 +1271,7 @@
         }
       }
 
-      Object.keys(settings).forEach(function(key) {
+      Object.keys(settings).forEach(function (key) {
         checkIFrame(key)
       })
     }
@@ -1278,7 +1337,7 @@
       )
     }
 
-    Object.keys(settings).forEach(function(iframeId) {
+    Object.keys(settings).forEach(function (iframeId) {
       if (isIFrameResizeEnabled(iframeId)) {
         trigger(eventName, event, settings[iframeId].iframe, iframeId)
       }
@@ -1288,7 +1347,7 @@
   function setupEventListeners() {
     addEventListener(window, 'message', iFrameListener)
 
-    addEventListener(window, 'resize', function() {
+    addEventListener(window, 'resize', function () {
       resizeIFrames('resize')
     })
 
@@ -1364,9 +1423,7 @@
           setupIFrame(element, options)
         }
 
-        return this.filter('iframe')
-          .each(init)
-          .end()
+        return this.filter('iframe').each(init).end()
       }
     }
   }