+++ src/iframeResizer.js
@@ -46,6 +46,7 @@
       maxWidth: Infinity,
       minHeight: 0,
       minWidth: 0,
+      mouseEvents: true,
       resizeFrom: 'parent',
       scrolling: false,
       sizeHeight: true,
@@ -61,6 +62,8 @@
       onMessage: function () {
         warn('onMessage function not defined')
       },
+      onMouseEnter: function () {},
+      onMouseLeave: function () {},
       onResized: function () {},
       onScroll: function () {
         return true
@@ -158,7 +161,7 @@
 
     function processMsg() {
       var data = msg.substr(msgIdLen).split(':')
-      var height = data[1] ? Number.parseInt(data[1], 10) : 0
+      var height = data[1] ? parseInt(data[1], 10) : 0
       var iframe = settings[data[0]] && settings[data[0]].iframe
       var compStyle = getComputedStyle(iframe)
 
@@ -175,11 +178,9 @@
       if (compStyle.boxSizing !== 'border-box') {
         return 0
       }
-      var top = compStyle.paddingTop
-        ? Number.parseInt(compStyle.paddingTop, 10)
-        : 0
+      var top = compStyle.paddingTop ? parseInt(compStyle.paddingTop, 10) : 0
       var bot = compStyle.paddingBottom
-        ? Number.parseInt(compStyle.paddingBottom, 10)
+        ? parseInt(compStyle.paddingBottom, 10)
         : 0
       return top + bot
     }
@@ -189,10 +190,10 @@
         return 0
       }
       var top = compStyle.borderTopWidth
-        ? Number.parseInt(compStyle.borderTopWidth, 10)
+        ? parseInt(compStyle.borderTopWidth, 10)
         : 0
       var bot = compStyle.borderBottomWidth
-        ? Number.parseInt(compStyle.borderBottomWidth, 10)
+        ? parseInt(compStyle.borderBottomWidth, 10)
         : 0
       return top + bot
     }
@@ -298,10 +299,12 @@
           msgBody +
           '}'
       )
+
       on('onMessage', {
         iframe: messageData.iframe,
         message: JSON.parse(msgBody)
       })
+
       log(iframeId, '--')
     }
 
@@ -320,11 +323,8 @@
           document.documentElement.clientWidth,
           window.innerWidth || 0
         ),
-        offsetTop: Number.parseInt(iFramePosition.top - bodyPosition.top, 10),
-        offsetLeft: Number.parseInt(
-          iFramePosition.left - bodyPosition.left,
-          10
-        ),
+        offsetTop: parseInt(iFramePosition.top - bodyPosition.top, 10),
+        offsetLeft: parseInt(iFramePosition.left - bodyPosition.left, 10),
         scrollTop: window.pageYOffset,
         scrollLeft: window.pageXOffset,
         documentHeight: document.documentElement.clientHeight,
@@ -510,6 +510,30 @@
       }
     }
 
+    function onMouse(event) {
+      var mousePos = {}
+
+      if (Number(messageData.width) === 0 && Number(messageData.height) === 0) {
+        var data = getMsgBody(9).split(':')
+        mousePos = {
+          x: data[1],
+          y: data[0]
+        }
+      } else {
+        mousePos = {
+          x: messageData.width,
+          y: messageData.height
+        }
+      }
+
+      on(event, {
+        iframe: messageData.iframe,
+        screenX: Number(mousePos.x),
+        screenY: Number(mousePos.y),
+        type: messageData.type
+      })
+    }
+
     function on(funcName, val) {
       return chkEvent(iframeId, funcName, val)
     }
@@ -524,6 +548,14 @@
 
         case 'message':
           forwardMsgFromIFrame(getMsgBody(6))
+          break
+
+        case 'mouseenter':
+          onMouse('onMouseEnter')
+          break
+
+        case 'mouseleave':
+          onMouse('onMouseLeave')
           break
 
         case 'autoResize':
@@ -564,7 +596,19 @@
           break
 
         default:
-          resizeIFrame()
+          if (
+            Number(messageData.width) === 0 &&
+            Number(messageData.height) === 0
+          ) {
+            warn(
+              'Unsupported message received (' +
+                messageData.type +
+                '), this is likely due to the iframe containing a later ' +
+                'version of iframe-resizer than the parent page'
+            )
+          } else {
+            resizeIFrame()
+          }
       }
     }
 
@@ -886,7 +930,9 @@
       ':' +
       settings[iframeId].resizeFrom +
       ':' +
-      settings[iframeId].widthCalculationMethod
+      settings[iframeId].widthCalculationMethod +
+      ':' +
+      settings[iframeId].mouseEvents
     )
   }